Extract out sensor enums file
